Make Responders handle panics.
Trying to handle a panic outside genhttp was problematic, as the defer
writing the response would already be called. Defers get called on a
last-in-first-out basis, and genhttp.Handle is always the last thing to
run before handler code runs, so its Responder.Write defer is always the
first non-handler defer called.

This is a problem for recovering from panics, because it means each
handler would need to recover from panics itself. Further, the Response
would often try to set an HTTP status code, which would probably be set
to 0 because there was an unexpected panic during the handling. And
writing a 0 status code causes a panic, itself. Masking the original
panic. This is very annoying and hard to debug.

By making Responders responsible for defining what happens to the
Response in the face of an error, we can define just once what happens
when a panic occurs, and what gets written back to the client.

As a drawback, any panics that happen before the response gets created
are handled by the net/http panic handler. Which is a bit of a bummer.